Rename commands.py to core.py in discord.commands
