+++ discord/commands/__init__.py
@@ -23,7 +23,7 @@ FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 DEALINGS IN THE SOFTWARE.
 """
 
-from .commands import *
 from .context import *
+from .core import *
 from .errors import *
 from .permissions import *
